Ask EU citizens 'When are you coming to the UK?' for clear results
Adds a new question for EU citizens: 'When are you coming to the
UK?'. If they answer 'before 1 January 2021', take them to an
outcome that says they do not need a visa. We could have used the
existing outcome_no_visa_needed.erb template here but we may want
to write something bespoke and having a separate template allows
us to do that more easily.

If citizens answer 'from 1 January 2021', route them to the
existing flow for non-visa nationals, for example Argentina:
https://www.gov.uk/check-uk-visa/y/argentina

This commit also removes a failing Croatia test which adds no
value.

Trello: https://trello.com/c/WoFhTMCi/2196-5-1-dec-logic-changes-for-eu-citizens-check-if-you-need-a-uk-visa